moving check_name_args to main
