+++ src/utils/utils.py
@@ -23,7 +23,6 @@ from torch.autograd import Variable
 from torch.nn import Dropout, Linear, Parameter, init
 
 from .config import Params
-from .. import tasks as task_modules
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -477,33 +476,3 @@ class MaskedMultiHeadSelfAttention(Seq2SeqEncoder):
 
 def assert_for_log(condition, error_message):
     assert condition, error_message
-
-
-def check_arg_name(args):
-    """ Raise error if obsolete arg names are present. """
-    # Mapping - key: old name, value: new name
-    name_dict = {
-        "task_patience": "lr_patience",
-        "do_train": "do_pretrain",
-        "train_for_eval": "do_target_task_training",
-        "do_eval": "do_full_eval",
-        "train_tasks": "pretrain_tasks",
-        "eval_tasks": "target_tasks",
-        "eval_data_fraction": "target_train_data_fraction",
-        "eval_val_interval": "target_train_val_interval",
-        "eval_max_vals": "target_train_max_vals",
-        "load_eval_checkpoint": "load_target_train_checkpoint",
-        "eval_data_fraction": "target_train_data_fraction",
-    }
-    for task in task_modules.ALL_GLUE_TASKS + task_modules.ALL_SUPERGLUE_TASKS:
-        assert_for_log(
-            not args.regex_contains("^{}_".format(task)),
-            "Error: Attempting to load old task-specific args for task %s, please refer to the master branch's default configs for the most recent task specific argument structures."
-            % task,
-        )
-    for old_name, new_name in name_dict.items():
-        assert_for_log(
-            old_name not in args,
-            "Error: Attempting to load old arg name [%s], please update to new name [%s]"
-            % (old_name, name_dict[old_name]),
-        )
